% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bootPower}
\alias{bootPower}
\title{Perform bootstrap power analysis on a list of square distance matrices.}
\usage{
bootPower(dm_list, boot_number = 100, subjects_per_group = 10,
  alpha = 0.05)
}
\arguments{
\item{dm_list}{a list of square distance matrices, with names}

\item{boot_number}{the number of bootstrap samples to perform on each distance matrix in the list}

\item{subjects_per_group}{number of subjects per group to sample.}

\item{alpha}{the threshold for PERMANOVA type I error}
}
\value{
A data frame relating PERMANOVA power to effect size quantified by the coefficient of determination (R^2) and omega-squared.
}
\description{
Estimates the statistical power of PERMANOVA testing to detect the group-level effect in the given distance matrices, based upon bootstrap sampling.
}
\examples{
bootPower(lapply(simPower(),calcUJstudy))
}
\seealso{
\code{\link{simPower}}, \code{\link{bootDM}}
}

