% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calcR2}
\alias{calcR2}
\title{Calculate the coefficient of determination (R-squared).}
\usage{
calcR2(perm = PERMANOVA(dm))
}
\arguments{
\item{perm}{output from the \code{\link{PERMANOVA}} function}
}
\value{
A numeric value.
}
\description{
The proportion of distance accounted for by the grouping factor.
}
\details{
A wrapper for the \code{\link{PERMANOVA}} function, which itself utilizes the \code{\link{adonis}} function included in the \code{\link{vegan}} package.
}
\examples{
calcR2(calcUJstudy(simStudy()))
}
\seealso{
\code{\link{calcUJstudy}}, \code{\link{calcWJstudy}}
}

