% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calcUJsample}
\alias{calcUJsample}
\title{Calculate unweighted Jaccard distance between two OTU vectors}
\usage{
calcUJsample(sampleA = rareSamp(simSamp(), 0.5),
  sampleB = rareSamp(simSamp(), 0.5))
}
\arguments{
\item{sampleA}{numeric vector labelled with OTU names}

\item{sampleB}{numeric vector labelled with OTU names}
}
\value{
numeric unweighted Jaccard distance
}
\description{
Wrapper for \code{\link{vegan}} package distance function, to calculate unweighted Jaccard distance.
}
\examples{
calcUJsample(rareSamp(simSamp(100),0.5),rareSamp(simSamp(100),0.5))
}
\seealso{
\code{\link{simSamp}}, \code{\link{rareSamp}}
}

