% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calcWJsample}
\alias{calcWJsample}
\title{Calculate weighted Jaccard distance between two OTU vectors}
\usage{
calcWJsample(sampleA = rareSamp(simSamp(, 10), 0.5),
  sampleB = rareSamp(simSamp(, 10), 0.5))
}
\arguments{
\item{sampleA}{numeric vector labelled with OTU names}

\item{sampleB}{numeric vector labelled with OTU names}
}
\value{
numeric weighted Jaccard distance
}
\description{
Wrapper for \code{\link{vegan}} package distance function, to calculate weighted Jaccard distance.
}
\examples{
calcWJsample(rareSamp(simSamp(100,10),0.5),rareSamp(simSamp(100,10),0.5))
}
\seealso{
\code{\link{simSamp}}, \code{\link{rareSamp}}
}

