% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/micropower.R
\name{estimatePower}
\alias{estimatePower}
\title{Perform bootstrap power analysis
Estimates the statistical power of PERMANOVA testing to detect the group-level effect for a given set of parameters matrices, based upon bootstrap sampling.}
\usage{
estimatePower(group_sizes, boot_args = list(), distance_calc = calcUJstudy,
  ...)
}
\arguments{
\item{group_sizes}{numeric vector representing subjects per exposure/intervention group}

\item{boot_args}{a list with named elements specifying additional arguments to \code{\link{bootPower}}}

\item{distance_calc}{distance function to be used to calculation (either \code{\link{calcUJstudy}} or \code{\link{calcWJstudy}})}

\item{...}{additional arguments to \code{\link{simPower}}}
}
\value{
A data frame with columns effect and power giving the estimted power to detect an effect of a given size
}
\description{
Perform bootstrap power analysis
Estimates the statistical power of PERMANOVA testing to detect the group-level effect for a given set of parameters matrices, based upon bootstrap sampling.
}
\examples{
calculatePower(c(10,10),otu_number=100,rare_depth=.25,effect_range=seq(0,.3,length.out=10))
}
\seealso{
\code{\link{simPower}}, \code{\link{bootDM}}
}

