% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/micropower.R
\name{groupNames}
\alias{groupNames}
\title{Convert subject names to group-level names.}
\usage{
groupNames(x)
}
\arguments{
\item{x}{character vector of names to be converted to groups}
}
\value{
character vector of group names
}
\description{
Convert subject names to group-level names to permit PERMANOVA testing and analysis of group-level effects.
}
\details{
Input distance matrix colnames and rownames must be formatted "g2s12" for group 2, subject 12.
}
\examples{
groupNames(rownames(calcUJstudy(simStudy())))
}
\seealso{
\code{\link{calcUJstudy}}, \code{\link{calcWJstudy}}
}

