% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hashMean}
\alias{hashMean}
\title{Simulate a list of OTU tables to create a hash table relating depth of subsampling (rarefying) to mean pairwise distance}
\usage{
hashMean(rare_levels = runif(1000, 0, 1), rep_per_level = 1,
  otu_number = 1000, sequence_depth = 1)
}
\arguments{
\item{rare_levels}{the levels of random subsampling (rarefying), as the proportion of sequence counts to retain after subsampling, to be tested}

\item{rep_per_level}{number of different levels (proportion of sequence counts to retain after subsampling) tested}

\item{otu_number}{number of simulated OTUs}

\item{sequence_depth}{number of sequence counts per OTU bin}
}
\value{
list of two-dimensional-matrix OTU tables
}
\description{
Extends the \code{\link{simSamp}} function to generate a list of OTU tables representing different levels of subsampling.
}
\details{
For weighted analysis, specify a sequence depth greater than 1; for unweighted analyses, sequence depth is 1 (default).
}
\examples{
hashMean(,10,1000,10)
sapply(hashMean(runif(100,0,1),1,1000,10),FUN=function(x) {mean(lowerTriDM(calcWJstudy(x)))})
}
\seealso{
\code{\link{simSamp}}, \code{\link{calcWJstudy}}
}

