% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hashSD}
\alias{hashSD}
\title{Simulate a list of OTU tables to create a hash table relating OTU number to standard deviation of pairwise distances}
\usage{
hashSD(rare_depth = 0.5, otu_number_range = c(10, 100, 1000, 10000),
  sim_number = 100, sequence_depth = 10)
}
\arguments{
\item{rare_depth}{proportion of sequence counts to retain after subsampling}

\item{otu_number_range}{numeric vector, including the range of numbers of simulated OTUs to test}

\item{sim_number}{number of simulated subjects per OTU number}

\item{sequence_depth}{number of sequence counts per OTU bin}
}
\value{
list of two-dimensional-matrix OTU tables
}
\description{
Extends the \code{\link{simSamp}} function to generate a list of OTU tables representing different OTU numbers.
}
\details{
For weighted analysis, specify a sequence depth greater than 1; for unweighted analyses, sequence depth is 1 (default).
}
\examples{
hashSD(0.5,c(10,100,1000),1000,10)
sapply(hashSD(),FUN=function(x) {sd(lowerTriDM(calcWJstudy(x)))})
}
\seealso{
\code{\link{simSampW}}, \code{\link{calcDistW}}
}

