% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lowerTriDM}
\alias{lowerTriDM}
\title{Extract the pairwise distances from the lower triangle of a square distance matrix.}
\usage{
lowerTriDM(dm)
}
\arguments{
\item{dm}{a square distance matrix}
}
\value{
A vector of pairwise distances.
}
\description{
Produces a vector of pairwise distances from the lower triangle of a symmetric, square distance matrix.
}
\details{
A wrapper for the \code{\link{lower.tri}} function.
}
\examples{
lowerTriDM(calcUJstudy(simStudy()))
}
\seealso{
\code{\link{calcUJstudy}}, \code{\link{calcWJstudy}}, \code{\link{readDM}}
}

