% Generated by roxygen2 (4.0.1): do not edit by hand
\name{PERMANOVA}
\alias{PERMANOVA}
\title{Perform PERMANOVA testing.}
\usage{
PERMANOVA(dm)
}
\arguments{
\item{dm}{a square distance matrix}
}
\value{
Typical output for analysis of variance.
}
\description{
A wrapper for the \code{\link{adonis}} function included in the \code{\link{vegan}} package, restricted to a single grouping factor.
}
\details{
Does not permit specifying strata within which to constrain permutations; \code{\link{adonis}} should be used for this purpose.
}
\examples{
PERMANOVA(calcUJstudy(simStudy()))
}
\seealso{
\code{\link{calcUJstudy}}, \code{\link{calcWJstudy}}
}

