% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rareSamp}
\alias{rareSamp}
\title{Rarefy a simulated OTU vector}
\usage{
rareSamp(otu_vector, rare_depth = 0.5)
}
\arguments{
\item{otu_vector}{numeric vector labelled with OTU names}

\item{rare_depth}{proportion of sequence counts to retain after subsampling}
}
\value{
numeric vector labelled with OTU names
}
\description{
Performs random subsampling without replacement from simulated OTU vector.
}
\examples{
rareSamp(simSamp(100,10),0.6)
}
\seealso{
\code{\link{simSampU}}, \code{\link{simSampW}}
}

