% Generated by roxygen2 (4.0.1): do not edit by hand
\name{readDM}
\alias{readDM}
\title{Read QIIME-formatted distance matrices}
\usage{
readDM(filepath = file.choose())
}
\arguments{
\item{filepath}{the path to the tab-separated distance matrix file}
}
\value{
A square distance matrix.
}
\description{
Imports a tab-separated distance matrix file, as output by the QIIME pipeline.
}
\details{
A wrapper for the \code{\link{read.delim}} function, tailored to the output from QIIME's beta_diversity.py script.
}
\examples{
readDM()
}
\seealso{
\code{\link{writeOTUtable}}, \code{\link{readDMdir}}
}

