% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/micropower.R
\name{readDMdir}
\alias{readDMdir}
\title{Read all QIIME-formatted distance matrices in working directory}
\usage{
readDMdir(dir = ".")
}
\arguments{
\item{dir}{directory to list files from}
}
\value{
A list of square distance matrices.
}
\description{
Imports all tab-separated distance matrix files in the working directory, as output by the QIIME pipeline.
}
\details{
A wrapper for the \code{\link{read.delim}} function, tailored to the output from QIIME's beta_diversity.py script.
}
\examples{
\dontrun{readDMdir()}
}
\seealso{
\code{\link{writeOTUlist}}, \code{\link{readDM}}
}

