% Generated by roxygen2 (4.0.1): do not edit by hand
\name{simNull}
\alias{simNull}
\title{Simulate OTU table with null group-level effect for presence-absence or abundance-weighted analysis}
\usage{
simNull(group_size_vector = c(100, 100, 100), otu_number = 1000,
  sequence_depth = 1, rare_depth = 0.5)
}
\arguments{
\item{group_size_vector}{numeric vector representing subjects per exposure/intervention group}

\item{otu_number}{number of simulated OTUs}

\item{sequence_depth}{number of sequence counts per OTU bin}

\item{rare_depth}{proportion of sequence counts to retain after subsampling}
}
\value{
two-dimensional matrix OTU table, with row and column names to suit downstream analysis
}
\description{
Within-group distances simulated from OTU number and depth of subsampling (rarefying); no group-level differences.
}
\details{
For weighted analysis, specify a sequence depth greater than 1; for unweighted analyses, sequence depth is 1 (default).
}
\examples{
simNull(c(16,16,16),100,10,0.8)
}
\seealso{
\code{\link{simStudy}}, \code{\link{simPower}}
}

