% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/micropower.R
\name{simPower}
\alias{simPower}
\title{Simulate a list of OTU tables encoding a range of effect sizes for presence-absence or abundance-weighted analysis}
\usage{
simPower(group_size_vector = c(20, 25, 30), otu_number = 1000,
  sequence_depth = 1, rare_depth = 0.5, effect_range = seq(0, 0.3,
  length.out = 10))
}
\arguments{
\item{group_size_vector}{numeric vector representing subjects per exposure/intervention group. NOTE that this should be larger than the true group sizes if the output will be used for downstream power calculations. True group sizes multiplied by 20 provides a good compromise between efficiency and avoiding oversampling.}

\item{otu_number}{number of simulated OTUs}

\item{sequence_depth}{number of sequence counts per OTU bin}

\item{rare_depth}{proportion of sequence counts to retain after subsampling}

\item{effect_range}{range of proportions of unique community membership in affected group of subjects}
}
\value{
list of two-dimensional-matrix OTU tables, with row and column names to suit downstream analysis
}
\description{
Extends the \code{\link{simStudy}} function to generate a list of OTU tables according to a specified range of group-level effects.
}
\details{
For weighted analysis, specify a sequence depth greater than 1; for unweighted analyses, sequence depth is 1 (default).
}
\examples{
simPower(c(16,16,16),100,10,0.8)
sapply(simPower(c(16,16,16),100,10,0.8),
  FUN=function(x) {calcOmega2(calcWJstudy(x))})
}
\seealso{
\code{\link{simStudy}}, \code{\link{calcUJstudy}}, \code{\link{calcWJstudy}}
}

