% Generated by roxygen2 (4.0.1): do not edit by hand
\name{simStudy}
\alias{simStudy}
\title{Simulate OTU table for presence-absence or abundance-weighted analysis}
\usage{
simStudy(group_size_vector = c(100, 100, 100), otu_number = 1000,
  sequence_depth = 1, rare_depth = 0.4, effect = 0.2)
}
\arguments{
\item{group_size_vector}{numeric vector representing subjects per exposure/intervention group}

\item{otu_number}{number of simulated OTUs}

\item{sequence_depth}{number of sequence counts per OTU bin}

\item{rare_depth}{proportion of sequence counts to retain after subsampling}

\item{effect}{proportion of unique community membership in affected group of subjects}
}
\value{
two-dimensional matrix OTU table, with row and column names to suit downstream analysis
}
\description{
Incorporates within-group distance simulation from OTU number and depth of subsampling (rarefying) with group-level differences modeled by segregating OTU membership between groups.
}
\details{
For weighted analysis, specify a sequence depth greater than 1; for unweighted analyses, sequence depth is 1 (default).
}
\examples{
simStudy(c(16,16,16),100,10,0.8,0.1)
}
\seealso{
\code{\link{calcUJstudy}}, \code{\link{calcWJstudy}}
}

