% Generated by roxygen2 (4.0.1): do not edit by hand
\name{simTreeList}
\alias{simTreeList}
\title{Simulate a phylogenetic tree with specified OTU names}
\usage{
simTreeList(otu_table_list)
}
\arguments{
\item{otu_table_list}{a list of OTU table matrices with one column per subject and one row per OTU}
}
\value{
An object of the class \code{\link{phylo}}.
}
\description{
Generates a random phylogenetic tree based upon the OTUs included in a list of OTU tables.
}
\details{
A wrapper for the \code{\link{rtree}} function included with the \code{\link{ape}} package.
}
\examples{
simTreeList(simPower(c(100,100,100),100,10,0.5,seq(0,0.3,length.out=100)))
}
\seealso{
\code{\link{simPowerU}}, \code{\link{simPowerW}}, \code{\link{rtree}}, \code{\link{write.tree}}
}

