% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/micropower.R
\name{simTreeTable}
\alias{simTreeTable}
\title{Simulate a phylogenetic tree with specified OTU names}
\usage{
simTreeTable(otu_table)
}
\arguments{
\item{otu_table}{a matrix with one column per subject and one row per OTU}
}
\value{
An object of the class \code{\link{phylo}}.
}
\description{
Generates a random phylogenetic tree based upon the OTUs included in a single OTU table.
}
\details{
A wrapper for the \code{\link{rtree}} function included with the \code{\link{ape}} package.
}
\examples{
simTreeTable(simStudy(c(16,16,16),100,,0.8,0.1))
}
\seealso{
\code{\link{simStudy}}, \code{\link{rtree}}, \code{\link{write.tree}}
}

