% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/micropower.R
\name{writeOTUtable}
\alias{writeOTUtable}
\title{Export a simulated OTU table to permit analysis of non-Jaccard distances}
\usage{
writeOTUtable(otu_table, otu_table_name)
}
\arguments{
\item{otu_table}{a matrix with one column per subject and one row per OTU}

\item{otu_table_name}{file name to write to (.txt is appended automatically}
}
\value{
Writes a tab-seperated file to the working directory.
}
\description{
Exports a simulated OTU table as a tab-separated file to permit processing for pairwise distances by bioinformatic pipelines.
}
\examples{
writeOTUtable(simStudy(c(16,16,16),100,1,0.8,0.1),"otu_table_export")
}
\seealso{
\code{\link{simStudy}}, \code{\link{simStudy}}
}

