% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bootDM}
\alias{bootDM}
\title{Take a bootstrap sample from a square distance matrix.}
\usage{
bootDM(dm, subjects_per_group)
}
\arguments{
\item{dm}{a square distance matrix}

\item{subjects_per_group}{number of subjects per group to sample.}
}
\value{
A square distance matrix.
}
\description{
Random sample with replacement from a distance matrix, with resulting matrix specified by number of subjects per group.
}
\examples{
bootDM(calcUJstudy(simStudy()))
}
\seealso{
\code{\link{calcUJstudy}}, \code{\link{calcWJstudy}}
}

