% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calcWJstudy}
\alias{calcWJstudy}
\title{Calculate pairwise weighted Jaccard distances from an OTU table}
\usage{
calcWJstudy(otu_table)
}
\arguments{
\item{otu_table}{a matrix with one column per subject and one row per OTU}
}
\value{
square matrix of pairwise distances
}
\description{
Generates a square matrix of pairwise weighted Jaccard distances from an OTU table.
}
\examples{
calcWJstudy(simStudy(c(16,16,16),100,10,0.8,0.1))
}
\seealso{
\code{\link{simStudyU}}, \code{\link{calcDistW}}
}

