% Generated by roxygen2 (4.0.1): do not edit by hand
\name{groupNames}
\alias{groupNames}
\title{Convert distance matrix subject names to group-level names.}
\usage{
groupNames(dm)
}
\arguments{
\item{dm}{a square distance matrix}
}
\value{
A square distance matrix.
}
\description{
Convert distance matrix subject names to group-level names to permit PERMANOVA testing and analysis of group-level effects.
}
\details{
Input distance matrix colnames and rownames must be formatted "g2s12" for group 2, subject 12.
}
\examples{
groupNames(calcUJstudy(simStudy()))
}
\seealso{
\code{\link{calcUJstudy}}, \code{\link{calcWJstudy}}
}

