% Generated by roxygen2 (4.0.1): do not edit by hand
\name{readDMmetric}
\alias{readDMmetric}
\title{Read all QIIME-formatted distance matrices in working directory that are labelled with a specified distance metric}
\usage{
readDMmetric(metric = "weighted_normalized_unifrac")
}
\value{
A list of square distance matrices.
}
\description{
Imports all tab-separated distance matrix files in the working directory with names that include the specified distance metric, as output by the QIIME pipeline.
}
\details{
A wrapper for the \code{\link{read.delim}} function, tailored to the output from QIIME's beta_diversity.py script.
}
\examples{
readDMmetric("weighted_normalized_unifrac")
}
\seealso{
\code{\link{writeOTUlist}}, \code{\link{readDMdir}}
}

